move matrix() to tests
