+++ doc/code/qml_qaoa.rst
@@ -100,4 +100,3 @@ Cycle Optimization
     :no-heading:
     :no-inheritance-diagram:
     :no-inherited-members:
-    :skip: kron
